Refactored test code to hopefully resolve code climate issues
